Replace original NumberParagraphsTreeProcessor with the NumberSectionParagraphs variant.
